% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Build rOpenSci docs}
\usage{
build_site(repo_url, deploy_url = NULL)
}
\arguments{
\item{repo_url}{full URL of the git remote (used to customize the template)}

\item{deploy_url}{deprecated}
}
\description{
Builds custom rOpenSci pkgdown site from  a git remote, and store in
a zip file to be saved as a CI artifact.
}
\examples{
\dontrun{
build_site('https://github.com/ropensci/magick')
}
}
